+++ crates/ruma-client-api/src/session/sso_login_with_provider.rs
@@ -7,7 +7,7 @@ pub mod v3 {
     //!
     //! [spec]: https://spec.matrix.org/latest/client-server-api/#get_matrixclientv3loginssoredirectidpid
 
-    use http::header::LOCATION;
+    use http::header::{LOCATION, SET_COOKIE};
     use ruma_common::{
         api::{request, response, Metadata},
         metadata,
@@ -43,6 +43,10 @@ pub mod v3 {
         /// Redirect URL to the SSO identity provider.
         #[ruma_api(header = LOCATION)]
         pub location: String,
+
+        /// Cookie storing state to secure the SSO process.
+        #[ruma_api(header = SET_COOKIE)]
+        pub cookie: Option<String>,
     }
 
     impl Request {
@@ -55,7 +59,7 @@ pub mod v3 {
     impl Response {
         /// Creates a new `Response` with the given SSO URL.
         pub fn new(location: String) -> Self {
-            Self { location }
+            Self { location, cookie: None }
         }
     }