+++ crates/ruma-client-api/CHANGELOG.md
@@ -23,6 +23,7 @@ Improvements:
   login types.
 - Add deprecated `address` and `medium` 3PID fields for `m.login.password`
   login type.
+- Add optional cookie field to `session::sso_login*::v3` responses.
 
 # 0.17.4